Fixed functionality of GetLocations button. Refactored file names to Location App.
