+++ app/app/.settings/org.eclipse.buildship.core.prefs
@@ -2,7 +2,7 @@ arguments=
 auto.sync=false
 build.scans.enabled=false
 connection.gradle.distribution=GRADLE_DISTRIBUTION(VERSION(6.3))
-connection.project.dir=
+connection.project.dir=..
 eclipse.preferences.version=1
 gradle.user.home=
 java.home=C\:/Program Files/Java/jdk1.8.0_231
