+++ app/app/src/androidTest/java/com/example/locationapp/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.ticketer;
+package com.example.locationapp;
 
 import android.content.Context;
