+++ app/app/src/main/java/com/example/locationapp/LocationFragment.java
@@ -1,15 +1,14 @@
-package com.example.ticketer;
+package com.example.locationapp;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toast;
 
 import androidx.fragment.app.ListFragment;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
-import com.example.ticketer.Tasks.FetchLocationsTask;
+import com.example.locationapp.Tasks.FetchLocationsTask;
 
 public class LocationFragment extends ListFragment {
 
@@ -31,7 +30,7 @@ public class LocationFragment extends ListFragment {
         ((MainActivity)getActivity()).setFragmentRefreshListener(new MainActivity.FragmentRefreshListener() {
             @Override
             public void onRefresh() {
-                myToast();
+                refresh();
             }
         });
 
@@ -50,7 +49,4 @@ public class LocationFragment extends ListFragment {
         new FetchLocationsTask(this).execute();
     }
 
-    public void myToast() {
-        Toast.makeText(this.getContext(), "This is called from MainActivity", Toast.LENGTH_SHORT).show();
-    }
 }
