+++ app/app/src/main/java/com/example/locationapp/Tasks/UpdateLocationTask.java
@@ -1,35 +1,26 @@
-package com.example.ticketer.Tasks;
+package com.example.locationapp.Tasks;
 
 import android.annotation.SuppressLint;
-import android.app.Activity;
 import android.content.Context;
-import android.content.Intent;
 import android.os.AsyncTask;
 import android.os.Build;
 import android.util.Log;
 import android.widget.Toast;
 
 import androidx.annotation.RequiresApi;
-import androidx.core.app.ActivityCompat;
-
-import com.example.ticketer.LoginActivity;
-import com.example.ticketer.MainActivity;
 
 import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.io.BufferedReader;
-import java.io.BufferedWriter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.net.URLEncoder;
-import java.nio.charset.StandardCharsets;
 
 public class UpdateLocationTask extends AsyncTask<String, Void, Integer> {
     public static final String API_URL = "http://10.0.2.2:8080/Locationapp/api";
