+++ app/app/src/main/java/com/example/locationapp/Tasks/UserLogin.java
@@ -1,4 +1,4 @@
-package com.example.ticketer.Tasks;
+package com.example.locationapp.Tasks;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
@@ -12,7 +12,7 @@ import android.widget.Toast;
 import androidx.annotation.RequiresApi;
 import androidx.core.app.ActivityCompat;
 
-import com.example.ticketer.MainActivity;
+import com.example.locationapp.MainActivity;
 
 import org.json.JSONArray;
 import org.json.JSONException;
