--- app/app/src/main/java/com/example/ticketer/RegisterActivity.java
@@ -1,109 +1,0 @@
-package com.example.ticketer;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.text.TextUtils;
-import android.view.View;
-import android.view.View.OnClickListener;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.ActionBar;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.navigation.fragment.NavHostFragment;
-
-import com.example.ticketer.Tasks.RegisterTask;
-
-public class RegisterActivity extends AppCompatActivity {
-
-    // UI references.
-    private EditText mFirstNameView;
-    private EditText mLastNameView;
-    private EditText mEmailView;
-    private EditText mUsernameView;
-    private EditText mPasswordView;
-    EditText passingUsername;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_register);
-
-        ActionBar actionBar = getSupportActionBar();
-        actionBar.setDisplayHomeAsUpEnabled(true);
-        actionBar.setTitle("Register");
-
-        mFirstNameView = findViewById(R.id.firstName);
-        mLastNameView = findViewById(R.id.lastName);
-        mEmailView = findViewById(R.id.email);
-        mUsernameView = findViewById(R.id.username);
-        mPasswordView = findViewById(R.id.password);
-
-        passingUsername = (EditText)findViewById(R.id.username);
-
-        Button mRegisterBtn = findViewById(R.id.register_btn);
-
-            mRegisterBtn.setOnClickListener(new OnClickListener() {
-                @Override
-                public void onClick(View view) {
-                    Intent intentMessage = new Intent(RegisterActivity.this, LoginActivity.class);
-                    String message = passingUsername.getText().toString();
-                    intentMessage.putExtra("username_key", message);
-                    attemptRegister();
-                }
-            });
-    }
-
-    private void attemptRegister() {
-        mFirstNameView.setError(null);
-        mLastNameView.setError(null);
-        mEmailView.setError(null);
-        mUsernameView.setError(null);
-        mPasswordView.setError(null);
-
-        String firstName = mFirstNameView.getText().toString();
-        String lastName = mLastNameView.getText().toString();
-        String email = mEmailView.getText().toString().toLowerCase();
-        String username = mUsernameView.getText().toString().toLowerCase();
-        String password = mPasswordView.getText().toString();
-
-        boolean cancel = false;
-        View focusView = null;
-
-        // validate info
-        if (TextUtils.isEmpty(password)) {
-            mPasswordView.setError(getString(R.string.error_field_required));
-            focusView = mPasswordView;
-            cancel = true;
-        }
-        if (TextUtils.isEmpty(username)) {
-            mUsernameView.setError(getString(R.string.error_field_required));
-            focusView = mUsernameView;
-            cancel = true;
-        }
-        if (TextUtils.isEmpty(email)) {
-            mEmailView.setError(getString(R.string.error_field_required));
-            focusView = mEmailView;
-            cancel = true;
-        }
-        if (TextUtils.isEmpty(lastName)) {
-            mLastNameView.setError(getString(R.string.error_field_required));
-            focusView = mLastNameView;
-            cancel = true;
-        }
-        if (TextUtils.isEmpty(firstName)) {
-            mFirstNameView.setError(getString(R.string.error_field_required));
-            focusView = mFirstNameView;
-            cancel = true;
-        }
-
-        if (cancel) {
-            focusView.requestFocus();
-        } else {
-            RegisterTask registerTask = new RegisterTask(this);
-            registerTask.execute(firstName, lastName, email, username, password);
-        }
-    }
-}
