+++ app/app/src/main/res/layout/activity_login.xml
@@ -64,19 +64,19 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/userName" />
 
-    <Button
-        android:id="@+id/register_button"
-        android:layout_width="118dp"
-        android:layout_height="56dp"
-        android:layout_marginStart="155dp"
-        android:layout_marginLeft="155dp"
-        android:layout_marginTop="44dp"
-        android:layout_marginEnd="156dp"
-        android:layout_marginRight="156dp"
-        android:layout_marginBottom="132dp"
-        android:text="@string/register_button"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/button_first" />
+<!--    <Button-->
+<!--        android:id="@+id/register_button"-->
+<!--        android:layout_width="118dp"-->
+<!--        android:layout_height="56dp"-->
+<!--        android:layout_marginStart="155dp"-->
+<!--        android:layout_marginLeft="155dp"-->
+<!--        android:layout_marginTop="44dp"-->
+<!--        android:layout_marginEnd="156dp"-->
+<!--        android:layout_marginRight="156dp"-->
+<!--        android:layout_marginBottom="132dp"-->
+<!--        android:text="@string/register_button"-->
+<!--        app:layout_constraintBottom_toBottomOf="parent"-->
+<!--        app:layout_constraintEnd_toEndOf="parent"-->
+<!--        app:layout_constraintStart_toStartOf="parent"-->
+<!--        app:layout_constraintTop_toBottomOf="@+id/button_first" />-->
 </androidx.constraintlayout.widget.ConstraintLayout>
