+++ app/app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Ticketer</string>
+    <string name="app_name">Location App</string>
     <string name="action_settings">Settings</string>
     <!-- Strings used for fragments for navigation -->
     <string name="first_fragment_label">First Fragment</string>
@@ -21,19 +21,13 @@
     <string name="prompt_email">Email</string>
     <string name="prompt_username">Username</string>
     <string name="prompt_password">Password</string>
-    <string name="action_register">Register</string>
-    <string name="action_register_short">Register</string>
-    <string name="enter_title_of_ticket">enter title of ticket</string>
     <string name="submit">submit</string>
     <string name="logout">logout</string>
-    <string name="enter_ticket_description">enter ticket description</string>
     <string name="main_fragment_label">main fragment label</string>
     <string name="register_fragment_label">register fragment label</string>
-    <string name="ticket_fragment_label">ticket fragment label</string>
     <string name="Login_fragment_label">login fragment label</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
-    <string name="register_button">Create Account</string>
     <string name="updateLocation">Update your Location</string>
     <string name="getLocation">Get a list of users latest location.</string>
     <string name="logged_in_as">Logged in as:</string>
