+++ app/app/src/test/java/com/example/locationapp/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.ticketer;
+package com.example.locationapp;
 
 import org.junit.Test;
